legal(icon): Set spdx for the project branding
Sets the repository to have an spdx file for the icon.svg (from ack file)

This is intended to standardize a lot of the acknowledgements & licensing components of the repository. The spdx convention allows for better automation, and keeping it consistent avoids having to implement custom parsers/data models for licensing.